+++ package.json
@@ -10,8 +10,7 @@
     "prepublish": "npm run build",
     "pretty-quick": "pretty-quick",
     "publish": "node tools/npm-publish.js",
-    "test": "lerna run test --stream && npm run coverage",
-    "updateTestArtifacts": "node tools/updateTestArtifacts.js"
+    "test": "lerna run test --stream && npm run coverage"
   },
   "husky": {
     "hooks": {
