+++ packages/target-decisioning-engine/src/geoProvider.spec.js
@@ -6,7 +6,7 @@ import {
 } from "./geoProvider";
 import { HTTP_HEADER_FORWARDED_FOR } from "./constants";
 
-const ARTIFACT_BLANK = require("../test/test-artifacts/TEST_ARTIFACT_BLANK.json");
+const ARTIFACT_BLANK = require("../test/schema/artifacts/TEST_ARTIFACT_BLANK.json");
 
 require("jest-fetch-mock").enableMocks();
