+++ packages/target-decisioning-engine/test/schema/artifacts/TEST_ARTIFACT_AB_SIMPLE.json
@@ -1,6 +1,5 @@
 {
-  "**TEST--targetAdminUrls**": [
-    "https://experience.adobe.com/#/@adobesummit2018/target/activities/activitydetails/A-B/simple_ab_expendable-mbox",
+  "targetAdminActivityUrls": [
     "https://experience.adobe.com/#/@adobesummit2018/target/activities/activitydetails/A-B/superfluous_ab_json"
   ],
   "version": "1.0.0",
@@ -143,80 +142,6 @@
                 "type": "json",
                 "eventToken": "abzfLHwlBDBNtz9ALey2fJNWHtnQtQrJfmRrQugEa2qCnQ9Y9OaLL2gsdrWQTvE54PwSz67rmXWmSnkXpSSS2Q==",
                 "content": { "doMagic": false, "importantValue": 75 }
-              }
-            ],
-            "metrics": []
-          }
-        }
-      ],
-      "expendable-mbox": [
-        {
-          "ruleKey": "334640",
-          "activityId": 334640,
-          "meta": {
-            "activity.id": 334640,
-            "activity.name": "simple ab expendable-mbox",
-            "activity.type": "ab",
-            "experience.id": 0,
-            "experience.name": "Experience A",
-            "location.name": "expendable-mbox",
-            "location.type": "mbox",
-            "location.id": 0,
-            "audience.ids": [],
-            "offer.id": 632233,
-            "offer.name": "/simple_ab_expendable-mbox/experiences/0/pages/0/zones/0/1582825446969",
-            "option.id": 2,
-            "option.name": "Offer2"
-          },
-          "condition": {
-            "and": [
-              { "<=": [0, { "var": "allocation" }] },
-              { ">=": [50, { "var": "allocation" }] }
-            ]
-          },
-          "consequence": {
-            "name": "expendable-mbox",
-            "options": [
-              {
-                "type": "html",
-                "eventToken": "YPoPRnGmGGit+QxnpxYFjWqipfsIHvVzTQxHolz2IpSCnQ9Y9OaLL2gsdrWQTvE54PwSz67rmXWmSnkXpSSS2Q==",
-                "content": "<div>hello</div>"
-              }
-            ],
-            "metrics": []
-          }
-        },
-        {
-          "ruleKey": "334640",
-          "activityId": 334640,
-          "meta": {
-            "activity.id": 334640,
-            "activity.name": "simple ab expendable-mbox",
-            "activity.type": "ab",
-            "experience.id": 1,
-            "experience.name": "Experience B",
-            "location.name": "expendable-mbox",
-            "location.type": "mbox",
-            "location.id": 0,
-            "audience.ids": [],
-            "offer.id": 632234,
-            "offer.name": "/simple_ab_expendable-mbox/experiences/1/pages/0/zones/0/1582825446975",
-            "option.id": 3,
-            "option.name": "Offer3"
-          },
-          "condition": {
-            "and": [
-              { "<": [50, { "var": "allocation" }] },
-              { ">=": [100, { "var": "allocation" }] }
-            ]
-          },
-          "consequence": {
-            "name": "expendable-mbox",
-            "options": [
-              {
-                "type": "html",
-                "eventToken": "YPoPRnGmGGit+QxnpxYFjZNWHtnQtQrJfmRrQugEa2qCnQ9Y9OaLL2gsdrWQTvE54PwSz67rmXWmSnkXpSSS2Q==",
-                "content": "<div>goodbye</div>"
               }
             ],
             "metrics": []
