+++ packages/target-decisioning-engine/test/schema/artifacts/TEST_ARTIFACT_GEO.json
@@ -1,5 +1,5 @@
 {
-  "**TEST--targetAdminUrls**": [
+  "targetAdminActivityUrls": [
     "https://experience.adobe.com/#/@adobesummit2018/target/activities/activitydetails/A-B/greg_geo_test_a_bjun0220201434"
   ],
   "version": "1.0.0",
