+++ packages/target-decisioning-engine/test/schema/artifacts/TEST_ARTIFACT_GLOBAL_MBOX.json
@@ -1,5 +1,5 @@
 {
-  "**TEST--targetAdminUrls**": [
+  "targetAdminActivityUrls": [
     "https://experience.adobe.com/#/@adobesummit2018/target/activities/activitydetails/Experience-Targeting/global_mbox_browserhtml",
     "https://experience.adobe.com/#/@adobesummit2018/target/activities/activitydetails/A-B/ab_tres_experienceglobalmbox",
     "https://experience.adobe.com/#/@adobesummit2018/target/activities/activitydetails/A-B/global_ab_html"
