+++ packages/target-decisioning-engine/test/schema/artifacts/TEST_ARTIFACT_PAGELOAD_VEC_XT.json
@@ -1,5 +1,5 @@
 {
-  "**TEST--targetAdminUrls**": [
+  "targetAdminActivityUrls": [
     "https://experience.adobe.com/#/@adobesummit2018/target/activities/activitydetails/Experience-Targeting/xt_vec_non-spa_correctparamslocal-target-test"
   ],
   "version": "1.0.0",
