+++ packages/target-decisioning-engine/test/schema/artifacts/TEST_ARTIFACT_PARAMS.json
@@ -1,5 +1,5 @@
 {
-  "**TEST--targetAdminUrls**": [
+  "targetAdminActivityUrls": [
     "https://experience.adobe.com/#/@adobesummit2018/target/activities/activitydetails/A-B/redundant-mbox"
   ],
   "version": "1.0.0",
