--- packages/target-decisioning-engine/test/test-artifacts/TEST_ARTIFACT_ADDRESS.json
@@ -1,164 +1,0 @@
-{
-  "**TEST--targetAdminUrls**": [
-    "https://experience.adobe.com/#/@adobesummit2018/target/activities/activitydetails/A-B/form_based_activity-offer2-feb1920201034"
-  ],
-  "version": "1.0.0",
-  "meta": { "clientCode": "adobesummit2018", "environment": "production" },
-  "globalMbox": "target-global-mbox",
-  "geoTargetingEnabled": true,
-  "responseTokens": [
-    "activity.id",
-    "activity.name",
-    "experience.id",
-    "experience.name",
-    "experience.trafficAllocationId",
-    "experience.trafficAllocationType",
-    "geo.city",
-    "geo.country",
-    "geo.state",
-    "offer.id",
-    "offer.name",
-    "option.id",
-    "option.name"
-  ],
-  "remoteMboxes": ["target-global-mbox"],
-  "remoteViews": ["home"],
-  "localMboxes": [
-    "alloy-test-scope-1",
-    "bannedtest",
-    "blippi",
-    "boogie",
-    "browser-mbox",
-    "composite2",
-    "crazy2",
-    "daterange-mbox",
-    "dell-countdown-mbox",
-    "dell-greeting-mbox",
-    "demo-engineering-flags",
-    "demo-geo-offer1",
-    "demo-marketing-flags",
-    "demo-marketing-offer1",
-    "demo-marketing-offer2",
-    "disney",
-    "dwalling-myFlags",
-    "expendable-mbox",
-    "foo",
-    "geo",
-    "gregtest42",
-    "gw-amp-email",
-    "gw-amp-email-thumbnail",
-    "jason-flags",
-    "kitty",
-    "local123456",
-    "location",
-    "macros",
-    "multinot",
-    "offer1",
-    "offer2",
-    "offer3",
-    "onboarding-mbox-1",
-    "perf1",
-    "product",
-    "redundant-mbox",
-    "sam-mbox",
-    "sam-mbox-2",
-    "something",
-    "specialk",
-    "splendid-mbox",
-    "superfluous-mbox",
-    "target-global-mbox",
-    "testoffer",
-    "testoffer2",
-    "vanilla-mbox"
-  ],
-  "localViews": ["about", "contact"],
-  "rules": {
-    "mboxes": {
-      "offer2": [
-        {
-          "ruleKey": "333312",
-          "activityId": 333312,
-          "meta": {
-            "activity.id": 333312,
-            "activity.name": "Form Based Activity - offer2 - Feb 19 2020, 10:34",
-            "activity.type": "ab",
-            "experience.id": 0,
-            "experience.name": "Experience A",
-            "location.name": "offer2",
-            "location.type": "mbox",
-            "location.id": 0,
-            "audience.ids": [5356811],
-            "offer.id": 630815,
-            "offer.name": "/form_based_activity-offer2-feb1920201034/experiences/0/pages/0/zones/0/1582137291528",
-            "option.id": 2,
-            "option.name": "Offer2"
-          },
-          "condition": {
-            "and": [
-              {
-                "and": [
-                  { "<=": [0, { "var": "allocation" }] },
-                  { ">=": [50, { "var": "allocation" }] }
-                ]
-              },
-              { "in": ["bar", { "var": "page.url_lc" }] }
-            ]
-          },
-          "consequence": {
-            "name": "offer2",
-            "options": [
-              {
-                "type": "json",
-                "eventToken": "mWtD0yDAXMnesyQOa7/jS2qipfsIHvVzTQxHolz2IpSCnQ9Y9OaLL2gsdrWQTvE54PwSz67rmXWmSnkXpSSS2Q==",
-                "content": { "baz": 1 }
-              }
-            ],
-            "metrics": []
-          }
-        },
-        {
-          "ruleKey": "333312",
-          "activityId": 333312,
-          "meta": {
-            "activity.id": 333312,
-            "activity.name": "Form Based Activity - offer2 - Feb 19 2020, 10:34",
-            "activity.type": "ab",
-            "experience.id": 1,
-            "experience.name": "Experience B",
-            "location.name": "offer2",
-            "location.type": "mbox",
-            "location.id": 0,
-            "audience.ids": [5356811],
-            "offer.id": 630814,
-            "offer.name": "/form_based_activity-offer2-feb1920201034/experiences/1/pages/0/zones/0/1582137291524",
-            "option.id": 3,
-            "option.name": "Offer3"
-          },
-          "condition": {
-            "and": [
-              {
-                "and": [
-                  { "<": [50, { "var": "allocation" }] },
-                  { ">=": [100, { "var": "allocation" }] }
-                ]
-              },
-              { "in": ["bar", { "var": "page.url_lc" }] }
-            ]
-          },
-          "consequence": {
-            "name": "offer2",
-            "options": [
-              {
-                "type": "json",
-                "eventToken": "mWtD0yDAXMnesyQOa7/jS5NWHtnQtQrJfmRrQugEa2qCnQ9Y9OaLL2gsdrWQTvE54PwSz67rmXWmSnkXpSSS2Q==",
-                "content": { "baz": 2 }
-              }
-            ],
-            "metrics": []
-          }
-        }
-      ]
-    },
-    "views": {}
-  }
-}
