+++ packages/target-nodejs-sdk/packfiles.txt
@@ -1,1 +1,1 @@
-CHANGELOG.md CODE_OF_CONDUCT.md LICENSE LICENSE.txt README.md
+CHANGELOG.md CODE_OF_CONDUCT.md LICENSE README.md
